Add auth, metrics, blocklist and WebAuthn modules
